requirements.txt
transformers==4.8.0
torch==1.8.1
torchtext==0.6.0
numpy==1.18.5
tqdm==4.56.0
regex==2020.11.13
GitPython==3.1.14
